+++ CHANGELOG.md
@@ -6,3 +6,9 @@ All notable changes to this project will be documented in this file.
 - Added csv export
 - Added constraints for positions in same team
 - Changed constraint setting interface
+
+## [v1.3]
+- Fixed bug with setting lineup positions
+
+## [v1.4]
+- Added min salary constraint.